ATRFS initial commit
Support added for:
  Atari DOS 1
  Atari DOS 2.0s
  Atari DOS 2.5
  MyDOS 4.53
And implicit support for Atari DOS 2.0d, as it behaves exactly the
same as MyDOS on a 720-sector image.

This has had minimal testing, so there may well be bugs.

The framework is in place for adding more file systems, and I plan
to add SpartaDOS next.